// src/model/Response/User/UserPostRes.tsx
export interface UserPostRes {
    UID:               number;
    image_user:        string;
    name_user:         string;
    nickname_user:     string;
    province:          string;
    gmail_user:        string;
    password_user:     string;
    phone:             string;
    facebook:          string;
    lineID:            string;
    datetime_register: Date;
    type_user:         number;
    typename_user:     string;
}